refactor: 마이페이지 관련 API 정리
