+++ src/main/java/com/example/wegobe/like/controller/LikeController.java
@@ -1,7 +1,5 @@
 package com.example.wegobe.like.controller;
 
-import com.example.wegobe.config.SecurityUtil;
-import com.example.wegobe.gathering.dto.response.GatheringResponseDto;
 import com.example.wegobe.like.dto.LikeResponseDto;
 import com.example.wegobe.like.service.LikeService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -11,7 +9,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
 
 @Tag(name = "Like", description = "동행 찜 API")
 @SecurityRequirement(name = "Bearer Authentication")
@@ -40,14 +37,4 @@ public class LikeController {
         return ResponseEntity.ok(likeService.isLiked(gatheringId));
     }
 
-    /**
-     * 특정 유저의 찜 목록 조회
-     */
-    @Operation(summary = "내 찜 목록 조회", description = "현재 사용자의 찜한 동행 목록을 조회합니다.")
-    @GetMapping("/my")
-    public ResponseEntity<List<GatheringResponseDto>> getLikedGatherings() {
-        Long kakaoId = SecurityUtil.getCurrentKakaoId();
-        List<GatheringResponseDto> likedGatherings = likeService.getLikedGatherings(kakaoId);
-        return ResponseEntity.ok(likedGatherings);
-    }
 }
